Added class and test for validation class and sorted the appropriate tests to the cooresponding classes.
